+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `place_perp_order_with_max_ts`
 - Fix precision loss in `ZoDexMarket::lots_to_price`
 - Add `EventFillLog` and `OtcFill` events
+- Implement `Display` for `Symbol`
 - BREAKING: Change accounts passed to `cache_oracle`
 - BREAKING: Change accounts passed to `withdraw`
 - BREAKING: Bump `anchor-lang` to `v0.25.0`